requirements.txt
python-dotenv>=1.0.0
requests>=2.20.0
flask>=2.0.0
pytest>=7.0.0
streamlit>=1.20.0
Pillow>=9.0.0
google-generativeai>=0.8.5